backend/requirements.txt
flask
flask-cors
requests
python-dotenv
PyJWT
google-auth
google-generativeai
Pillow
beautifulsoup4
boto3
google-cloud-storage
